pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amx"
version = "0.1.0"
description = "Agentic Metadata Extractor — AI-powered CLI to infer and manage database metadata"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Omer Yasir Kucuk"},
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pandas>=2.0",
    "chromadb>=0.5",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-text-splitters>=0.3",
    "tiktoken>=0.7",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.7",
    "litellm>=1.40",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "unstructured>=0.15",
    "python-magic>=0.4",
    "pypdf>=4.0",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "gitpython>=3.1",
    "boto3>=1.34",
    "requests>=2.31",
]

[project.scripts]
amx = "amx.cli:main"

[tool.setuptools.packages.find]
include = ["amx*"]
